Database.HDBC

Description

Welcome to HDBC, the Haskell Database Connectivity library.

Written by John Goerzen, jgoerzen@complete.org

Introduction

Welcome to HDBC, Haskell Database Connectivity.

HDBC provides an abstraction layer between Haskell programs and SQL relational databases. This lets you write database code once, in Haskell, and have it work with any number of backend SQL databases (MySQL, Oracle, PostgreSQL, ODBC-compliant databases, etc.)

HDBC is modeled loosely on Perl's DBI interface http://search.cpan.org/~timb/DBI/DBI.pm, though it has also been influenced by Python's DB-API v2, JDBC in Java, and HSQL in Haskell.

HDBC is a from-scratch effort. It is not a reimplementation of HSQL, though its purpose is the same.

Features

Features of HDBC include:

Ability to use replacable parameters to let one query be executed multiple times (eliminates the need for an escape function)

Ability to access returned rows by column number

Ability to read data from the SQL server on-demand rather than reading the entire result set up front

HUnit testsuite for each backend driver

Well-defined standard API and easy backend driver implementation

Lazy reading of the entire result set (think hGetContents, but for the results of SELECT) (see sFetchAllRows)

Support for translation between Haskell and SQL types

Support for querying database server properties

Add-on package (hdbc-missingh) to integrate with MissingH, providing a database backend for AnyDBM.

Support for querying metadata such as column names.

Support for querying additional metadata (column types, etc.)

Available Drivers

Here is a list of known drivers as of January 26, 2009:

Sqlite v3

Available from http://software.complete.org/hdbc-sqlite3. Or, to participate in development, use git clone git://git.complete.org/hdbc-sqlite3

PostgreSQL

Available from http://software.complete.org/hdbc-postgresql. Or, to participate in development, use git clone git://git.complete.org/hdbc-postgresql

ODBC

Available from http://software.complete.org/hdbc-odbc. Or, to partitipace in development, use git clone git://git.complete.org/hdbc-odbc

MySQL

MySQL users have two choices: the first is the ODBC driver, which works and has been tested against MySQL on both Linux/Unix and Windows platforms. There is also an alpha-quality native MySQL driver available for download at http://hackage.haskell.org/cgi-bin/hackage-scripts/package/HDBC-mysql with a homepage at http://www.maubi.net/~waterson/hacks/hdbc-mysql.html.

In addition, there is one integration package: hdbc-anydbm. This integrates with the AnyDBM library http://software.complete.org/anydbm. It lets any HDBC database act as a backend for the AnyDBM interface. Available from http://software.complete.org/hdbc-anydbm. Or, to participate in development, use darcs get --partial http://darcs.complete.org/hdbc-anydbm

The latest version of HDBC itself is available from http://software.complete.org/hdbc. Or, to participate in development, use git clone git://git.complete.org/hdbc.

Typing of transfer data

data SqlValue

SqlValue is he main type for expressing Haskell values to SQL databases.

INTRODUCTION TO SQLVALUE

This type is used to marshall Haskell data to and from database APIs. HDBC driver interfaces will do their best to use the most accurate and efficient way to send a particular value to the database server.

Values read back from the server are constructed with the most appropriate SqlValue constructor. fromSql or safeFromSql can then be used to convert them into whatever type is needed locally in Haskell.

Most people will use toSql and fromSql instead of manipulating SqlValues directly.

EASY CONVERSIONS BETWEEN HASKELL TYPES

Conversions are powerful; for instance, you can call fromSql on a SqlInt32 and get a String or a Double out of it. This class attempts to Do The Right Thing whenever possible, and will raise an error when asked to do something incorrect. In particular, when converting to any type except a Maybe, SqlNull as the input will cause an error to be raised.

Conversions are implemented in terms of the Data.Convertible module, part of the convertible package. You can refer to its documentation, and import that module, if you wish to parse the Left result from safeFromSql yourself, or write your own conversion instances.

Here are some notes about conversion:

Fractions of a second are not preserved on time values

There is no safeToSql because toSql never fails.

See also toSql, safeFromSql, fromSql, nToSql, iToSql, posixToSql.

ERROR CONDITIONS

There may sometimes be an error during conversion. For instance, if you have a SqlString and are attempting to convert it to an Integer, but it doesn't parse as an Integer, you will get an error. This will be indicated as an exception if using fromSql, or a Left result if using safeFromSql.

SPECIAL NOTE ON POSIXTIME

Note that a NominalDiffTime or POSIXTime is converted to SqlDiffTime by toSql. HDBC cannot differentiate between NominalDiffTime and POSIXTime since they are the same underlying type. You must construct SqlPOSIXTime manually or via posixToSql, or use SqlUTCTime.

DETAILS ON SQL TYPES

HDBC database backends are expected to marshal date and time data back and forth using the appropriate representation for the underlying database engine. Databases such as PostgreSQL with builtin date and time types should see automatic conversion between these Haskell types to database types. Other databases will be presented with an integer or a string. Care should be taken to use the same type on the Haskell side as you use on the database side. For instance, if your database type lacks timezone information, you ought not to use ZonedTime, but instead LocalTime or UTCTime. Database type systems are not always as rich as Haskell. For instance, for data stored in a TIMESTAMP WITHOUT TIME ZONE column, HDBC may not be able to tell if it is intended as UTCTime or LocalTime data, and will happily convert it to both, upon your request. It is your responsibility to ensure that you treat timezone issues with due care.

This behavior also exists for other types. For instance, many databases do not have a Rational type, so they will just use the show function and store a Rational as a string.

The conversion between Haskell types and database types is complex, and generic code in HDBC or its backends cannot possibly accomodate every possible situation. In some cases, you may be best served by converting your Haskell type to a String, and passing that to the database.

UNICODE AND BYTESTRINGS

Beginning with HDBC v2.0, interactions with a database are presumed to occur in UTF-8.

To accomplish this, whenever a ByteString must be converted to or from a String, the ByteString is assumed to be in UTF-8 encoding, and will be decoded or encoded as appropriate. Database drivers will generally present text or string data they have received from the database as a SqlValue holding a ByteString, which fromSql will automatically convert to a String, and thus automatically decode UTF-8, when you need it. In the other direction, database drivers will generally convert a SqlString to a ByteString in UTF-8 encoding before passing it to the database engine.

If you are handling some sort of binary data that is not in UTF-8, you can of course work with the ByteString directly, which will bypass any conversion.

Due to lack of support by database engines, lazy ByteStrings are not passed to database drivers. When you use toSql on a lazy ByteString, it will be converted to a strict ByteString for storage. Similarly, fromSql will convert a strict ByteString to a lazy ByteString if you demand it.

EQUALITY OF SQLVALUE

Two SqlValues are considered to be equal if one of these hold. The first comparison that can be made is controlling; if none of these comparisons can be made, then they are not equal:

Both are NULL

Both represent the same type and the encapsulated values are considered equal by applying (==) to them

The values of each, when converted to a string, are equal.

STRING VERSIONS OF TIMES

Default string representations are given as comments below where such are non-obvious. These are used for fromSql when a String is desired. They are also defaults for representing data to SQL backends, though individual backends may override them when a different format is demanded by the underlying database. Date and time formats use ISO8601 date format, with HH:MM:SS added for time, and -HHMM added for timezone offsets.

DEPRECATED CONSTRUCTORS

SqlEpochTime and SqlTimeDiff are no longer created automatically by any toSql or fromSql functions or database backends. They may still be manually constructed, but are expected to be removed in a future version. Although these two constructures will be removed, support for marshalling to and from the old System.Time data will be maintained as long as System.Time is, simply using the newer data types for conversion.

Constructors

SqlString String

SqlByteString ByteString

SqlWord32 Word32

SqlWord64 Word64

SqlInt32 Int32

SqlInt64 Int64

SqlInteger Integer

SqlChar Char

SqlBool Bool

SqlDouble Double

SqlRational Rational

SqlLocalDate Day
Local YYYY-MM-DD (no timezone)

SqlLocalTimeOfDay TimeOfDay
Local HH:MM:SS (no timezone)

SqlZonedLocalTimeOfDay TimeOfDay TimeZone
Local HH:MM:SS -HHMM. Converts to and from (TimeOfDay, TimeZone).

SqlLocalTime LocalTime
Local YYYY-MM-DD HH:MM:SS (no timezone)

SqlZonedTime ZonedTime
Local YYYY-MM-DD HH:MM:SS -HHMM. Considered equal if both convert to the same UTC time.

SqlUTCTime UTCTime
UTC YYYY-MM-DD HH:MM:SS

SqlDiffTime NominalDiffTime
Calendar diff between seconds. Rendered as Integer when converted to String, but greater precision may be preserved for other types or to underlying database.

SqlPOSIXTime POSIXTime
Time as seconds since midnight Jan 1 1970 UTC. Integer rendering as for SqlDiffTime.

SqlEpochTime Integer
DEPRECATED Representation of ClockTime or CalendarTime. Use SqlPOSIXTime instead.

SqlTimeDiff Integer
DEPRECATED Representation of TimeDiff. Use SqlDiffTime instead.

SqlNull
NULL in SQL or Nothing in Haskell

toSql :: Convertible a SqlValue => a -> SqlValue

Convert a value to an SqlValue. This function is simply a restricted-type wrapper around convert. See extended notes on SqlValue.

fromSql :: Convertible SqlValue a => SqlValue -> a

Convert from an SqlValue to a Haskell value. Any problem is indicated by calling error. This function is simply a restricted-type wrapper around convert. See extended notes on SqlValue.

safeFromSql :: Convertible SqlValue a => SqlValue -> ConvertResult a

Conversions to and from SqlValues and standard Haskell types.

This function converts from an SqlValue to a Haskell value. Many people will use the simpler fromSql instead. This function is simply a restricted-type wrapper around safeConvert.

nToSql :: Integral a => a -> SqlValue

Converts any Integral type to a SqlValue by using toInteger.

iToSql :: Int -> SqlValue

Convenience function for using numeric literals in your program.

posixToSql :: POSIXTime -> SqlValue

Convenience function for converting POSIXTime to a SqlValue, because toSql cannot do the correct thing in this instance.

Database Connections

class IConnection conn where

Main database handle object.

An IConnection object is created by specific functions in the module for an individual database. That is, the connect function -- which creates this object -- is not standardized through the HDBC interface.

A connection is closed by a call to disconnect.

A call to commit is required to make sure that your changes get committed to the database. In other words, HDBC has no support for autocommit, which we consider an outdated notion.

Minimal complete definition

disconnect, commit, rollback, run, prepare, clone, hdbcDriverName, hdbcClientVer, proxiedClientName, proxiedClientVer, dbServerVer, dbTransactionSupport, getTables, describeTable

Methods

disconnect :: conn -> IO ()

Disconnect from the remote database.

You do not need to explicitly close an IConnection object, but you may do so if you so desire. If you don't, the object will disconnect from the database in a sane way when it is garbage-collected. However, a disconnection may raise an error, so you are encouraged to explicitly call disconnect. Also, garbage collection may not run when the program terminates, and some databases really like an explicit disconnect.

So, bottom line is, you're best off calling disconnect directly, but the world won't end if you forget.

This function discards any data not committed already. Database driver implementators should explicitly call rollback if their databases don't do this automatically on disconnect.

Bad Things (TM) could happen if you call this while you have Statements active. In more precise language, the results in such situations are undefined and vary by database. So don't do it.

commit :: conn -> IO ()

Commit any pending data to the database.

Required to make any changes take effect.

rollback :: conn -> IO ()

Roll back to the state the database was in prior to the last commit or rollback.

runRaw :: conn -> String -> IO ()

Execute an SQL string, which may contain multiple queries. This is intended for situations where you need to run DML or DDL queries and aren't interested in results.

run :: conn -> String -> [SqlValue] -> IO Integer

Execute a single SQL query. Returns the number of rows modified (see execute for details). The second parameter is a list of replacement values, if any.

prepare :: conn -> String -> IO Statement

Prepares a statement for execution.

Question marks in the statement will be replaced by positional parameters in a later call to execute.

Please note that, depending on the database and the driver, errors in your SQL may be raised either here or by execute. Make sure you handle exceptions both places if necessary.

clone :: conn -> IO conn

Create a new Connection object, pointed at the same server as this object is. This will generally establish a separate physical connection.

When you wish to establish multiple connections to a single server, the correct way to do so is to establish the first connection with the driver-specific connection function, and then clone it for each additional connection.

This can be important when a database doesn't provide much thread support itself, and the HDBC driver module must serialize access to a particular database.

This can also be a handy utility function whenever you need a separate connection to whatever database you are connected to already.

hdbcDriverName :: conn -> String

The name of the HDBC driver module for this connection. Ideally would be the same as the database name portion of the Cabal package name. For instance, "sqlite3" or "odbc". This is the layer that is bound most tightly to HDBC.

hdbcClientVer :: conn -> String

The version of the C (or whatever) client library that the HDBC driver module is bound to. The meaning of this is driver-specific. For an ODBC or similar proxying driver, this should be the version of the ODBC library, not the eventual DB client driver.

proxiedClientName :: conn -> String

In the case of a system such as ODBC, the name of the database client/server in use, if available. For others, identical to hdbcDriverName.

proxiedClientVer :: conn -> String

In the case of a system such as ODBC, the version of the database client in use, if available. For others, identical to hdbcClientVer. This is the next layer out past the HDBC driver.

dbServerVer :: conn -> String

The version of the database server, if available.

dbTransactionSupport :: conn -> Bool

Whether or not the current database supports transactions. If False, then commit and rollback should be expected to raise errors.

MySQL is the only commonly-used database that is known to not support transactions entirely. Please see the MySQL notes in the ODBC driver for more information.

getTables :: conn -> IO [String]

The names of all tables accessible by the current connection, excluding special meta-tables (system tables).

You should expect this to be returned in the same manner as a result from fetchAllRows'.

All results should be converted to lowercase for you before you see them.

describeTable :: conn -> String -> IO [(String, SqlColDesc)]

Obtain information about the columns in a specific table. The String in the result set is the column name.

You should expect this to be returned in the same manner as a result from fetchAllRows'.

All results should be converted to lowercase for you before you see them.

Wrapped Connections

data ConnWrapper

Sometimes, it is annoying to use typeclasses with Haskell's type system. In those situations, you can use a ConnWrapper. You can create one with:

let wrapped = ConnWrapper iconn

You can then use this directly, since a ConnWrapper is also an IConnection. However, you will not be able to use private database functions on it.

Or, you can use withWConn.

Constructors

forall conn . IConnection conn => ConnWrapper conn

withWConn :: forall b. ConnWrapper -> (forall conn. IConnection conn => conn -> b) -> b

Unwrap a ConnWrapper and pass the embedded IConnection to a function. Example:

withWConn wrapped run $ "SELECT * from foo where bar = 1" []

Preparing Queries

sRun :: IConnection conn => conn -> String -> [Maybe String] -> IO Integer

Like run, but take a list of Maybe Strings instead of SqlValues.

quickQuery' :: IConnection conn => conn -> String -> [SqlValue] -> IO [[SqlValue]]

Strict version of quickQuery.

quickQuery :: IConnection conn => conn -> String -> [SqlValue] -> IO [[SqlValue]]

A quick way to do a query. Similar to preparing, executing, and then calling fetchAllRows on a statement. See also quickQuery'

Transaction Handling

This section concerns itself with writing (updating) a database.

In HDBC, as with many RDBMS implementations, every write to the database occurs within a transaction. No changes are visible (outside the current transaction) until a commit operation occurs, in which case all changes since the transaction started are atomically committed. Also, there is a rollback operation that can undo all changes since the transaction started.

HDBC does everything within a transaction. A transaction is implicitly entered when a connection to a database is established, and a transaction is implicitly entered after each call to commit or rollback as well.

The practical effect of this is that you must call commit after making changes to a database in order for those changes to become visible. You don't have to call commit after every change, just after a batch of them.

(Exceptions exist for databases that don't offer a high level of transaction isolation; but you should always play it safe and commit anyway.)

Database developers will also be experienced with the atomicity benefits of transactions, an explanation of which is outside the scope of this manual.

Errors occuring at the database level can leave a transaction in an indeterminate state, depending on the database. Some databases will refuse all queries until the next commit or rollback. The safe thing to do is to issue a commit or rollback after trapping any SqlError. Alternatively, you could use withTransaction, which will automatically handle this detail for you.

withTransaction :: IConnection conn => conn -> (conn -> IO a) -> IO a

Execute some code. If any uncaught exception occurs, run rollback and re-raise it. Otherwise, run commit and return.

This function, therefore, encapsulates the logical property that a transaction is all about: all or nothing.

The IConnection object passed in is passed directly to the specified function as a convenience.

This function traps all uncaught exceptions, not just SqlErrors. Therefore, you will get a rollback for any exception that you don't handle. That's probably what you want anyway.

Since all operations in HDBC are done in a transaction, this function doesn't issue an explicit "begin" to the server. You should ideally have called commit or rollback before calling this function. If you haven't, this function will commit or rollback more than just the changes made in the included action.

If there was an error while running rollback, this error will not be reported since the original exception will be propogated back. (You'd probably like to know about the root cause for all of this anyway.) Feedback on this behavior is solicited.

Connection Inquiries

Statements

data Statement

Execution

execute :: Statement -> [SqlValue] -> IO Integer

Execute the prepared statement, passing in the given positional parameters (that should take the place of the question marks in the call to prepare).

For non-SELECT queries, the return value is the number of rows modified, if known. If no rows were modified, you get 0. If the value is unknown, you get -1. All current HDBC drivers support this function and should never return -1.

For SELECT queries, you will always get 0.

This function should automatically call finish() to finish the previous execution, if necessary.

executeRaw :: Statement -> IO ()

Execute the statement as-is, without supplying any positional parameters. This is intended for statements for which the results aren't interesting or present (e.g., DDL or DML commands). If your query contains placeholders, this will certainly fail; use execute instead.

sExecute :: Statement -> [Maybe String] -> IO Integer

Like execute, but take a list of Maybe Strings instead of SqlValues.

executeMany :: Statement -> [[SqlValue]] -> IO ()

Execute the query with many rows. The return value is the return value from the final row as if you had called execute on it.

Due to optimizations that are possible due to different databases and driver designs, this can often be significantly faster than using execute multiple times since queries need to be compiled only once.

This is most useful for non-SELECT statements.

sExecuteMany :: Statement -> [[Maybe String]] -> IO ()

Like executeMany, but take a list of Maybe Strings instead of SqlValues.

Fetching Results

fetchRow :: Statement -> IO (Maybe [SqlValue])

Fetches one row from the DB. Returns Nothing if there are no more rows. Will automatically call finish when the last row is read.

fetchRowAL :: Statement -> IO (Maybe [(String, SqlValue)])

Like fetchRow, but instead of returning a list, return an association list from column name to value.

The keys of the column names are lowercase versions of the data returned by getColumnNames. Please heed the warnings there. Additionally, results are undefined if multiple columns are returned with identical names.

fetchRowMap :: Statement -> IO (Maybe (Map String SqlValue))

Similar to fetchRowAL, but return a Map instead of an association list.

sFetchRow :: Statement -> IO (Maybe [Maybe String])

Like fetchRow, but return a list of Maybe Strings instead of SqlValues.

fetchAllRows :: Statement -> IO [[SqlValue]]

Lazily fetch all rows from an executed Statement.

You can think of this as hGetContents applied to a database result set.

The result of this is a lazy list, and each new row will be read, lazily, from the database as the list is processed.

When you have exhausted the list, the Statement will be finished.

Please note that the careless use of this function can lead to some unpleasant behavior. In particular, if you have not consumed the entire list, then attempt to finish or re-execute the statement, and then attempt to consume more elements from the list, the result will almost certainly not be what you want.

But then, similar caveats apply with hGetContents.

Bottom line: this is a very convenient abstraction; use it wisely.

Use fetchAllRows' if you need something that is strict, without all these caveats.

fetchAllRows' :: Statement -> IO [[SqlValue]]

Strict version of fetchAllRows. Does not have the side-effects of fetchAllRows, but forces the entire result set to be buffered in memory.

fetchAllRowsAL :: Statement -> IO [[(String, SqlValue)]]

Like fetchAllRows, but instead of returning a list for each row, return an association list for each row, from column name to value.

See fetchRowAL for more details.

fetchAllRowsAL' :: Statement -> IO [[(String, SqlValue)]]

Strict version of fetchAllRowsAL

fetchAllRowsMap :: Statement -> IO [Map String SqlValue]

Like fetchAllRowsAL, but return a list of Maps instead of a list of association lists.

fetchAllRowsMap' :: Statement -> IO [Map String SqlValue]

Strict version of fetchAllRowsMap

sFetchAllRows :: Statement -> IO [[Maybe String]]

Like fetchAllRows, but return Maybe Strings instead of SqlValues.

sFetchAllRows' :: Statement -> IO [[Maybe String]]

Strict version of sFetchAllRows.

getColumnNames :: Statement -> IO [String]

Returns a list of the column names in the result. For maximum portability, you should not assume that information is available until after an execute function has been run.

Information is returned here directly as returned by the underlying database layer. Note that different databases have different rules about capitalization of return values and about representation of names of columns that are not simple columns. For this reason, it is suggested that you treat this information for display purposes only. Failing that, you should convert to lower (or upper) case, and use AS clauses for anything other than simple columns.

A simple getColumnNames implementation could simply apply map fst to the return value of describeResult.

Statement Inquires

describeResult :: Statement -> IO [(String, SqlColDesc)]

Obtain information about the columns in the result set. Must be run only after execute. The String in the result set is the column name.

You should expect this to be returned in the same manner as a result from fetchAllRows'.

All results should be converted to lowercase for you before you see them.

Please see caveats under getColumnNames for information on the column name field here.

Miscellaneous

finish :: Statement -> IO ()

Abort a query in progress -- usually not needed.

originalQuery :: Statement -> String

The original query that this Statement was prepared with.

Exceptions

data SqlError

The main HDBC exception object. As much information as possible is passed from the database through to the application through this object.

Errors generated in the Haskell layer will have seNativeError set to -1.

Constructors

SqlError

Fields

seState :: String

seNativeError :: Int

seErrorMsg :: String

throwSqlError :: SqlError -> IO a

A utility function to throw a SqlError. The mechanics of throwing such a thing differ between GHC 6.8.x, Hugs, and GHC 6.10. This function takes care of the special cases to make it simpler.

With GHC 6.10, it is a type-restricted alias for throw. On all other systems, it is a type-restricted alias for throwDyn.

catchSql :: IO a -> (SqlError -> IO a) -> IO a

Execute the given IO action.

If it raises a SqlError, then execute the supplied handler and return its return value. Otherwise, proceed as normal.

handleSql :: (SqlError -> IO a) -> IO a -> IO a

Like catchSql, with the order of arguments reversed.

sqlExceptions :: SqlError -> Maybe SqlError

Given an Exception, return Just SqlError if it was an SqlError, or Nothing otherwise. Useful with functions like catchJust.

handleSqlError :: IO a -> IO a

Catches SqlErrors, and re-raises them as IO errors with fail. Useful if you don't care to catch SQL errors, but want to see a sane error message if one happens. One would often use this as a high-level wrapper around SQL calls.

Column Types

These are defined in Database.HDBC.ColTypes but are available to programs importing Database.HDBC by default as well. See Database.HDBC.ColTypes for documentation.

module Database.HDBC.ColTypes

Threading

FIXME: this is draft information

Thread support in a generalized interface such as HDBC can be complicated because support for threading varies across database interfaces.

However, applications using HDBC should be able to rely upon at least a few basic guarantees:

The HDBC modules may freely be imported and used across all threads.

HDBC modules may also freely share database connections and statements; the database or HDBC driver will be responsible for locking if necessary.

I use "share" in the same sense as Python's DB-API: multiple threads may use the resource without wrapping it in any lock.

However, there are some caveats to the above:

Not all databases support more than one active statement for a single connection. Therefore, for maximum portability, you should use a different connection to the database for each simultaneous query you wish to use. FIXME: describe when a statement is active.

Not all databases may support the level of multithreading described above. For those that don't, safe access will be restriced in the HDBC driver by using locks. Therefore, you can write portable code, but you only get real multithreading when databases really support it. Details of thread support should be documented in the HDBC driver for each specific database.